Add files via upload
Designed to be a quick review of using Scanner and PrintWriter objects, I took this assignment a step further and created a program that, given a year and a number, will return Billboard Hot 100 song from the specified year that placed at the specified number. It also allows the user to create his/her own playlist and search the song on Youtube if desired.